Made huge progress toward implementing 2+ multiplayer
Switched over player and player2 variables to dynamic variables that are
provided by the server. Multiple players show up on the screen now, but with
some glitches. Player1 is currently hard-coded because there is an error that
player1 is not defined b/c the script runs right past the listener and doesn't
receive the info in time. We will need to use promises or a callback to allow
the server to send the info before running the rest of the script.